+++ src/test/java/test/pivotal/pal/trackerapi/TimeEntryApiTest.java
@@ -27,7 +27,7 @@ public class TimeEntryApiTest {
 
     private final long projectId = 123L;
     private final long userId = 456L;
-    private TimeEntry timeEntry = new TimeEntry(projectId, userId, LocalDate.parse("2017-01-08"), 8L);
+    private TimeEntry timeEntry = new TimeEntry(projectId, userId, LocalDate.parse("2017-01-08"), 8);
 
     @Test
     public void testCreate() throws Exception {
@@ -41,7 +41,7 @@ public class TimeEntryApiTest {
         assertThat(createJson.read("$.projectId", Long.class)).isEqualTo(projectId);
         assertThat(createJson.read("$.userId", Long.class)).isEqualTo(userId);
         assertThat(createJson.read("$.date", String.class)).isEqualTo("2017-01-08");
-        assertThat(createJson.read("$.hours", Long.class)).isEqualTo(8);
+        assertThat(createJson.read("$.hours", Integer.class)).isEqualTo(8);
     }
 
     @Test
@@ -77,7 +77,7 @@ public class TimeEntryApiTest {
         assertThat(readJson.read("$.projectId", Long.class)).isEqualTo(projectId);
         assertThat(readJson.read("$.userId", Long.class)).isEqualTo(userId);
         assertThat(readJson.read("$.date", String.class)).isEqualTo("2017-01-08");
-        assertThat(readJson.read("$.hours", Long.class)).isEqualTo(8);
+        assertThat(readJson.read("$.hours", Integer.class)).isEqualTo(8);
     }
 
     @Test
@@ -85,7 +85,7 @@ public class TimeEntryApiTest {
         Long id = createTimeEntry();
         long projectId = 2L;
         long userId = 3L;
-        TimeEntry updatedTimeEntry = new TimeEntry(projectId, userId, LocalDate.parse("2017-01-09"), 9L);
+        TimeEntry updatedTimeEntry = new TimeEntry(projectId, userId, LocalDate.parse("2017-01-09"), 9);
 
 
         ResponseEntity<String> updateResponse = restTemplate.exchange("/time-entries/" + id, HttpMethod.PUT, new HttpEntity<>(updatedTimeEntry, null), String.class);
@@ -98,7 +98,7 @@ public class TimeEntryApiTest {
         assertThat(updateJson.read("$.projectId", Long.class)).isEqualTo(projectId);
         assertThat(updateJson.read("$.userId", Long.class)).isEqualTo(userId);
         assertThat(updateJson.read("$.date", String.class)).isEqualTo("2017-01-09");
-        assertThat(updateJson.read("$.hours", Long.class)).isEqualTo(9);
+        assertThat(updateJson.read("$.hours", Integer.class)).isEqualTo(9);
     }
 
     @Test
